fix: test email magic minting
