+++ packages/app/hooks/auth/use-magic-login.ts
@@ -1,5 +1,6 @@
 import { useCallback } from "react";
 
+import { useWeb3 } from "app/hooks/use-web3";
 import { BYPASS_EMAIL_WITH_INSECURE_KEYS } from "app/lib/constants";
 import { magic } from "app/lib/magic";
 import { overrideMagicInstance } from "app/utilities";
@@ -11,6 +12,7 @@ const LOGIN_MAGIC_ENDPOINT = "login_magic";
 export function useMagicLogin() {
   //#region hooks
   const { setAuthenticationStatus, login, logout } = useAuth();
+  const { setWeb3 } = useWeb3();
   //#endregion
 
   //#region methods
@@ -26,12 +28,18 @@ export function useMagicLogin() {
           did,
           phone_number: phoneNumber,
         });
+
+        const Web3Provider = (await import("@ethersproject/providers"))
+          .Web3Provider;
+        // @ts-ignore
+        const web3 = new Web3Provider(magic.rpcProvider);
+        setWeb3(web3);
       } catch (error) {
         logout();
         throw error;
       }
     },
-    [login, logout, setAuthenticationStatus]
+    [login, logout, setAuthenticationStatus, setWeb3]
   );
   const loginWithEmail = useCallback(
     async function loginWithEmail(email: string) {
@@ -50,12 +58,18 @@ export function useMagicLogin() {
           did,
           email: overrideEmail,
         });
+
+        const Web3Provider = (await import("@ethersproject/providers"))
+          .Web3Provider;
+        // @ts-ignore
+        const web3 = new Web3Provider(selectedMagicInstance.rpcProvider);
+        setWeb3(web3);
       } catch (error) {
         logout();
         throw error;
       }
     },
-    [login, logout, setAuthenticationStatus]
+    [login, logout, setAuthenticationStatus, setWeb3]
   );
   //#endregion
   return {
